Criada função de relatório + separado o código em arquivos
